moves src/app/state into src/state
