+++ stories/data-graphics/guides/SimpleAxisStory.svelte
@@ -1,5 +1,5 @@
 <script>
-import { responseToData } from '../../../src/app/state/store';
+import { responseToData } from '../../../src/state/store';
 import ACTIVE_TICKS_BUILD from '../../../tests/data/browser_engagement_active_ticks_build_id.json';
 
 const activeTicksBuild = ACTIVE_TICKS_BUILD.response;
